requirements.txt
# Flask and Web Server
flask>=2.3.0,<3.0.0
gunicorn>=21.0.0,<22.0.0

# Data Handling - Compatible versions for all Python 3.x
pandas>=2.0.0,<3.0.0
numpy>=1.21.0,<2.0.0
openpyxl>=3.1.0,<4.0.0
xlrd==2.0.1

# Google APIs
google-api-python-client>=2.100.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
gspread>=5.10.0,<6.0.0
google-generativeai>=0.3.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0
Pillow>=10.0.0,<11.0.0
beautifulsoup4>=4.12.0,<5.0.0
PyPDF2>=3.0.0,<4.0.0
plotly>=5.15.0,<6.0.0

# Additional dependencies for production
Werkzeug>=2.3.0,<3.0.0
Jinja2>=3.1.0,<4.0.0

# Add requests for health check
requests>=2.31.0,<3.0.0
